// src/api/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      incidents: {
        Row: {
          content_text: string | null;
          created_at: string;
          device_id: string;
          id: string;
          label: string;
          occurred_at: string;
          preview_url: string | null;
          resolved: boolean;
          risk: number;
          source_app: string | null;
          type: string;
        };
        Insert: {
          content_text?: string | null;
          created_at?: string;
          device_id: string;
          id?: string;
          label: string;
          occurred_at: string;
          preview_url?: string | null;
          resolved?: boolean;
          risk: number;
          source_app?: string | null;
          type: string;
        };
        Update: {
          content_text?: string | null;
          created_at?: string;
          device_id?: string;
          id?: string;
          label?: string;
          occurred_at?: string;
          preview_url?: string | null;
          resolved?: boolean;
          risk?: number;
          source_app?: string | null;
          type?: string;
        };
        Relationships: [
          {
            foreignKeyName: "incidents_device_id_fkey";
            columns: ["device_id"];
            isOneToOne: false;
            referencedRelation: "devices";
            referencedColumns: ["id"];
          }
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I;
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
  ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
      Insert: infer I;
    }
    ? I
    : never
  : never;